moved get_gsoc_projects to search project file
