+++ backend/apps/owasp/api/search/project.py
@@ -1,5 +1,7 @@
 """OWASP app project search API."""
 
+from functools import lru_cache
+
 from algoliasearch_django import raw_search
 
 from apps.owasp.models.project import Project
@@ -34,3 +36,23 @@ def get_projects(query, attributes=None, limit=25, page=1, searchable_attributes
         params["restrictSearchableAttributes"] = searchable_attributes
 
     return raw_search(Project, query, params)
+
+
+@lru_cache
+def get_gsoc_projects(year, attributes=None):
+    """Return GSoC projects from Algolia."""
+    from apps.owasp.models.project import Project
+
+    query = f"gsoc{year}"
+    searchable_attributes = ["idx_custom_tags", "idx_topics", "idx_tags"]
+
+    params = {
+        "attributesToHighlight": [],
+        "restrictSearchableAttributes": searchable_attributes,
+    }
+
+    if attributes:
+        params["attributesToRetrieve"] = attributes
+
+    results = raw_search(Project, query, params)
+    return results.get("hits", [])
